corrigida funcao atualizaCameraAcompanha
mudar solidSphere(int raio..) para solidSphere(double raio..)
mudar astro.raio_esf, astro.raio_trans, obs.xpos, obs.ypos, obs.zpos de float para double
corrigida funcao atualizaCameraAcompanha

+++ headers/funcoes_desenho.h
@@ -1,5 +1,5 @@
 //Funcao de desenhar uma esfera, retirada do exemplo de marte
-void solidSphere(int radius, int stacks, int columns){
+void solidSphere(double radius, int stacks, int columns){
     // cria uma quádrica
     GLUquadric* quadObj = gluNewQuadric();
     // estilo preenchido... poderia ser GLU_LINE, GLU_SILHOUETTE
